+++ jpl/rules/__init__.py
@@ -1,8 +1,8 @@
 from itertools import chain
 
-from src.jpl.rules.playbook import info, PlayBookExists
-from src.jpl.rules.pipeline import pipeline
-from src.jpl.rules.task import function, task
+from jpl.rules.playbook import info, PlayBookExists
+from jpl.rules.pipeline import pipeline
+from jpl.rules.task import function, task
 
 
 rules = list(chain.from_iterable([info.rules, pipeline.rules]))
