--- src/jpl/lint.py
@@ -1,7 +1,0 @@
-from src.jpl import JiggyPlaybookLint
-
-
-if __name__ == "__main__":
-    linter = JiggyPlaybookLint(
-        path="examples/jiggy-playbook-empty.yml", verbose=True, drop_passed=True
-    ).run()
